Initial implementation of new Dialog State API
