+++ compose/ui/ui/src/desktopMain/kotlin/androidx/compose/ui/awt/SwingDialog.desktop.kt
@@ -23,6 +23,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.rememberUpdatedState
+import androidx.compose.runtime.snapshots.Snapshot
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.input.key.KeyEvent
@@ -146,7 +147,12 @@ fun SwingDialog(
             // - Make the dialog displayable
             // - Size the dialog and the ComposeLayer correctly, so that we can draw it here
             if (!wasDisplayable && it.isDisplayable) {
-                it.renderImmediately()
+                Snapshot.withoutReadObservation {
+                    if (!it.isValid) {
+                        it.validate()
+                    }
+                    it.renderImmediately()
+                }
             }
         },
     )
