+++ compose/ui/ui/src/desktopMain/kotlin/androidx/compose/ui/awt/v2/SwingDialog.desktop.kt
@@ -26,8 +26,8 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.rememberUpdatedState
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.ExperimentalComposeUiApi
-import androidx.compose.ui.awt.ComposeWindow
-import androidx.compose.ui.awt.SwingWindow
+import androidx.compose.ui.awt.ComposeDialog
+import androidx.compose.ui.awt.SwingDialog
 import androidx.compose.ui.awt.toAwtRectangleSizeRoundedUp
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.input.key.KeyEvent
@@ -37,42 +37,39 @@ import androidx.compose.ui.util.componentListenerRef
 import androidx.compose.ui.util.setIcon
 import androidx.compose.ui.util.setUndecoratedSafely
 import androidx.compose.ui.util.windowListenerRef
-import androidx.compose.ui.util.windowStateListenerRef
-import androidx.compose.ui.window.FrameWindowScope
+import androidx.compose.ui.window.DialogWindowScope
+import androidx.compose.ui.window.LocalWindow
 import androidx.compose.ui.window.UndecoratedWindowDecoration
 import androidx.compose.ui.window.WindowDecoration
 import androidx.compose.ui.window.WindowLocationTracker
-import androidx.compose.ui.window.WindowPlacement
 import androidx.compose.ui.window.resizerThickness
-import androidx.compose.ui.window.roundToDimensionOrNull
 import androidx.compose.ui.window.toDpRect
+import androidx.compose.ui.window.v2.DialogState
 import androidx.compose.ui.window.v2.WindowBoundsProvider
 import androidx.compose.ui.window.v2.WindowGeometryProviderScope
-import androidx.compose.ui.window.v2.WindowScreenProvider
-import androidx.compose.ui.window.v2.WindowScreenProviderScope
-import androidx.compose.ui.window.v2.WindowSizeLimits
-import androidx.compose.ui.window.v2.WindowState
-import androidx.compose.ui.window.v2.rememberWindowState
-import java.awt.GraphicsDevice
-import java.awt.GraphicsEnvironment
+import androidx.compose.ui.window.v2.rememberDialogState
+import java.awt.Dialog.ModalityType
 import java.awt.event.ComponentAdapter
 import java.awt.event.ComponentEvent
 import java.awt.event.WindowAdapter
 import java.awt.event.WindowEvent
-import javax.swing.JFrame
+import javax.swing.JDialog
 import kotlinx.coroutines.isActive
 import kotlinx.coroutines.launch
 
+// TODO(demin): fix mouse hover after opening a dialog.
+//  When we open a modal dialog, ComposeLayer/mouseExited will
+//  never be called for the parent window. See ./gradlew run3
 
 /**
- * Similar to the corresponding [androidx.compose.ui.window.v2.Window] function, but additionally
- * allows configuring the underlying AWT window before it has been made displayable by providing an
- * [init] block.
+ * Similar to the corresponding [androidx.compose.ui.window.v2.DialogWindow] function, but
+ * additionally allows configuring the underlying AWT dialog before it has been made displayable,
+ * by providing an [init] block.
  *
  * This is useful to:
- * - Set window properties which cannot be changed after it has been made displayable, such as
+ * - Set dialog properties which cannot be changed after it has been made displayable, such as
  *   [java.awt.Window.setType].
- * - Adding listeners for events that can occur when the window becomes displayable/visible.
+ * - Adding listeners for events that can occur when the dialog becomes displayable/visible.
  *
  * IMPORTANT: this function should not be used to set properties which can be changed after the
  * window has been made displayable. Doing so can cause your code to stop working in the future if
@@ -82,22 +79,22 @@ import kotlinx.coroutines.launch
  *
  * To set these kinds of properties, use this pattern instead:
  * ```
- * Window( ... ) {
- *     // Window content here
+ * WindowDialog( ... ) {
+ *     // Dialog content here
  *     LaunchedEffect(window) {
- *         // Configure window here
+ *         // Configure dialog here
  *     }
  * }
  * ```
  *
- * @see androidx.compose.ui.window.v2.Window
+ * @see androidx.compose.ui.window.v2.DialogWindow
  */
 @ExperimentalComposeUiApi
 @Composable
 @ComposableOpenTarget(-1)
-fun SwingWindow(
+fun SwingDialog(
     onCloseRequest: () -> Unit,
-    state: WindowState = rememberWindowState(),
+    state: DialogState = rememberDialogState(),
     visible: Boolean = true,
     title: String = "Untitled",
     icon: Painter? = null,
@@ -107,12 +104,14 @@ fun SwingWindow(
     enabled: Boolean = true,
     focusable: Boolean = true,
     alwaysOnTop: Boolean = false,
-    sizeLimits: WindowSizeLimits = WindowSizeLimits.Unlimited,
-    onPreviewKeyEvent: (KeyEvent) -> Boolean = { false },
-    onKeyEvent: (KeyEvent) -> Boolean = { false },
-    init: (ComposeWindow) -> Unit,
-    content: @Composable FrameWindowScope.() -> Unit
+    onPreviewKeyEvent: ((KeyEvent) -> Boolean) = { false },
+    onKeyEvent: ((KeyEvent) -> Boolean) = { false },
+    modalityType: ModalityType = ModalityType.DOCUMENT_MODAL,
+    init: (ComposeDialog) -> Unit,
+    content: @Composable DialogWindowScope.() -> Unit
 ) {
+    val owner = LocalWindow.current
+
     val currentState by rememberUpdatedState(state)
     val currentTitle by rememberUpdatedState(title)
     val currentIcon by rememberUpdatedState(icon)
@@ -122,7 +121,7 @@ fun SwingWindow(
     val currentEnabled by rememberUpdatedState(enabled)
     val currentFocusable by rememberUpdatedState(focusable)
     val currentAlwaysOnTop by rememberUpdatedState(alwaysOnTop)
-    val currentSizeLimits by rememberUpdatedState(sizeLimits)
+    val currentModalityType by rememberUpdatedState(modalityType)
     val currentOnCloseRequest by rememberUpdatedState(onCloseRequest)
 
     val updater = remember(::ComponentUpdater)
@@ -130,12 +129,10 @@ fun SwingWindow(
     val listeners = remember {
         object {
             var windowListenerRef = windowListenerRef()
-            var windowStateListenerRef = windowStateListenerRef()
             var componentListenerRef = componentListenerRef()
 
-            fun removeFromAndClear(window: ComposeWindow) {
+            fun removeFromAndClear(window: ComposeDialog) {
                 windowListenerRef.unregisterFromAndClear(window)
-                windowStateListenerRef.unregisterFromAndClear(window)
                 componentListenerRef.unregisterFromAndClear(window)
             }
         }
@@ -143,61 +140,50 @@ fun SwingWindow(
 
     val coroutineContext = rememberCoroutineScope().coroutineContext
 
-    var window: ComposeWindow? by remember { mutableStateOf(null) }
-    SwingWindow(
+    var dialog: ComposeDialog? by remember { mutableStateOf(null) }
+    SwingDialog(
         visible = visible,
         onPreviewKeyEvent = onPreviewKeyEvent,
         onKeyEvent = onKeyEvent,
         create = {
-            val graphicsDevices = GraphicsEnvironment.getLocalGraphicsEnvironment().screenDevices
-            val currentDevice = currentState._screenId?.let { screenId ->
-                graphicsDevices.firstOrNull { it.iDstring == screenId }
+            val graphicsConfiguration = WindowLocationTracker.lastActiveGraphicsConfiguration
+            val dlg = if (owner != null) {
+                ComposeDialog(
+                    owner = owner,
+                    modalityType = currentModalityType,
+                    graphicsConfiguration = graphicsConfiguration,
+                    coroutineContext = coroutineContext
+                )
+            } else {
+                ComposeDialog(
+                    graphicsConfiguration = graphicsConfiguration,
+                    coroutineContext = coroutineContext
+                )
             }
-            val initialDevice = currentDevice
-                ?: state.screenRequests.tryReceive().getOrNull()?.getInitialScreenDevice()
-                ?: WindowScreenProvider.Default.getInitialScreenDevice()
-
-            val wnd = ComposeWindow(
-                graphicsConfiguration = initialDevice.defaultConfiguration,
-                coroutineContext = coroutineContext
-            )
-            // close state is controlled by WindowState.isOpen
-            wnd.defaultCloseOperation = JFrame.DO_NOTHING_ON_CLOSE
+            // close state is controlled by DialogState.isOpen
+            dlg.defaultCloseOperation = JDialog.DO_NOTHING_ON_CLOSE
             listeners.windowListenerRef.registerWithAndSet(
-                wnd,
+                dlg,
                 object : WindowAdapter() {
-                    override fun windowClosing(e: WindowEvent) {
+                    override fun windowClosing(e: WindowEvent?) {
                         currentOnCloseRequest()
                     }
                 }
             )
-            listeners.windowStateListenerRef.registerWithAndSet(wnd) {
-                currentState._placement = wnd.placement
-                currentState._isMinimized = wnd.isMinimized
-            }
             listeners.componentListenerRef.registerWithAndSet(
-                wnd,
+                dlg,
                 object : ComponentAdapter() {
                     fun applyBoundsChanges() {
-                        currentState._bounds = wnd.bounds.toDpRect()
-                        if (currentState._screenId != wnd.graphicsConfiguration.device.iDstring) {
-                            currentState._screenId = wnd.graphicsConfiguration.device.iDstring
-                        }
+                        currentState._bounds = dlg.bounds.toDpRect()
                     }
 
                     override fun componentShown(e: ComponentEvent) {
                         // Initialize all state properties
-                        currentState._placement = wnd.placement
-                        currentState._isMinimized = wnd.isMinimized
                         applyBoundsChanges()
                         currentState.isInitialized = true
                     }
 
                     override fun componentResized(e: ComponentEvent) {
-                        // we check placement here and in windowStateChanged,
-                        // because fullscreen changing doesn't
-                        // fire windowStateChanged, only componentResized
-                        currentState._placement = wnd.placement
                         applyBoundsChanges()
                     }
 
@@ -206,91 +192,58 @@ fun SwingWindow(
                     }
                 }
             )
-            WindowLocationTracker.onWindowCreated(wnd)
 
-            init(wnd)
-            window = wnd
+            init(dlg)
+            dialog = dlg
 
-            wnd
+            dlg
         },
         dispose = {
-            WindowLocationTracker.onWindowDisposed(it)
             // We need to remove them because AWT can still call them after dispose()
             listeners.removeFromAndClear(it)
             it.dispose()
         },
-        update = { window ->
-            if (!window.isDisplayable) {
-                window.initializePlacement(currentState)
-                window.initializeBounds(currentState)
+        update = { dialog ->
+            if (!dialog.isDisplayable) {
+                dialog.initializeBounds(currentState)
 
-                // Need to make the window displayable, to make awt.SwingWindow render the first
-                // frame before the window is visible.
+                // Need to make the dialog displayable, to make awt.SwingDialog render the first
+                // frame before the dialog is visible.
                 // Check window.isDisplayable again because initializeBounds could have already
                 // called pack(), and we don't need to do it twice
-                if (!window.isDisplayable) {
-                    window.preferredSize = window.size
-                    window.pack()  // Sizes to preferred size
+                if (!dialog.isDisplayable) {
+                    dialog.preferredSize = dialog.size
+                    dialog.pack()  // Sizes to preferred size
                 }
             }
 
             updater.update {
-                set(currentTitle, window::setTitle)
-                set(currentIcon, window::setIcon)
-                set(currentDecoration is UndecoratedWindowDecoration, window::setUndecoratedSafely)
-                set(currentTransparent, window::isTransparent::set)
-                set(currentResizable, window::setResizable)
-                set(currentEnabled, window::setEnabled)
-                set(currentFocusable, window::setFocusableWindowState)
-                set(currentAlwaysOnTop, window::setAlwaysOnTop)
-                set(currentSizeLimits.min) { window.minimumSize = it.roundToDimensionOrNull() }
-                set(currentSizeLimits.max) { window.maximumSize = it.roundToDimensionOrNull() }
-                set(currentDecoration.resizerThickness, window::undecoratedResizerThickness::set)
+                set(currentTitle, dialog::setTitle)
+                set(currentIcon, dialog::setIcon)
+                set(currentDecoration is UndecoratedWindowDecoration, dialog::setUndecoratedSafely)
+                set(currentTransparent, dialog::isTransparent::set)
+                set(currentResizable, dialog::setResizable)
+                set(currentEnabled, dialog::setEnabled)
+                set(currentFocusable, dialog::setFocusableWindowState)
+                set(currentAlwaysOnTop, dialog::setAlwaysOnTop)
+                set(currentModalityType, dialog::setModalityType)
+                set(currentDecoration.resizerThickness, dialog::undecoratedResizerThickness::set)
             }
         },
         content = content
     )
 
-    LaunchedEffect(window, state) {
-        val window = window ?: return@LaunchedEffect
+    LaunchedEffect(dialog, state) {
+        val dialog = dialog ?: return@LaunchedEffect
         launch {
             while (isActive) {
-                window.placement = state.placementRequests.receive()
-            }
-        }
-        launch {
-            while (isActive) {
-                window.isMinimized = state.isMinimizedRequests.receive()
-            }
-        }
-        launch {
-            while (isActive) {
-                window.setBoundsFrom(state.boundsRequests.receive())
+                dialog.setBoundsFrom(state.boundsRequests.receive())
             }
         }
     }
 }
 
-private fun WindowScreenProvider.getInitialScreenDevice(): GraphicsDevice {
-    val lastActiveConfig = WindowLocationTracker.lastActiveGraphicsConfiguration
-    val env = GraphicsEnvironment.getLocalGraphicsEnvironment()
-    val devices = env.screenDevices
-    val defaultDevice =
-        devices.firstOrNull { it.iDstring === lastActiveConfig?.device?.iDstring } ?:
-        env.defaultScreenDevice
-    return with(WindowScreenProviderScope(devices.toList(), defaultDevice)) {
-        getScreen().device
-    }
-}
-
-private fun ComposeWindow.initializePlacement(state: WindowState) {
-    val placementRequest = state.placementRequests.tryReceive().getOrNull()
-    val currentPlacement = state._placement
-
-    placement = placementRequest ?: currentPlacement ?: WindowPlacement.Floating
-}
-
-private fun ComposeWindow.initializeBounds(state: WindowState) {
+private fun ComposeDialog.initializeBounds(state: DialogState) {
     val boundsRequest = state.boundsRequests.tryReceive().getOrNull()
     val currentBounds = state._bounds
 
@@ -302,7 +255,7 @@ private fun ComposeWindow.initializeBounds(state: WindowState) {
     }
 }
 
-private fun ComposeWindow.setBoundsFrom(boundsProvider: WindowBoundsProvider) {
+private fun ComposeDialog.setBoundsFrom(boundsProvider: WindowBoundsProvider) {
     if (!isDisplayable) {
         // Give it a preferred size to avoid measuring via ComposeSceneMediator.preferredSize
         // when pack() is called
