+++ compose/ui/ui/src/desktopTest/kotlin/androidx/compose/ui/window/WindowTypingLocationTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 The Android Open Source Project
+ * Copyright 2026 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package androidx.compose.ui.window.window
+package androidx.compose.ui.window
 
 import androidx.compose.ui.focusedInputMethodRequests
 import androidx.compose.ui.sendCharTypedEvents
